+++ homebridge-ui/server.js
@@ -30,9 +30,6 @@ class PluginUiServer extends HomebridgePluginUiServer {
       tokensFile: tokensFile
     }
     const authentication = new Authentication(authConfig);
-    authentication.on('error', (error) => {
-;
-    })
 
     let data = {};
     switch (mode) {